修改权重为浮点类型

+++ gridfactorylib/src/main/java/com/toolsmi/gridfactory/GridItem.java
@@ -4,8 +4,8 @@ public class GridItem {
     int layout;
     int row;
     int column;
-    int rowWeight;
-    int columnWeight;
+    float rowWeight;
+    float columnWeight;
     int rowSpan;
     int columnSpan;
 
@@ -18,7 +18,7 @@ public class GridItem {
         column = Integer.parseInt(items[2]);
         rowSpan = Integer.parseInt(items[3]);
         columnSpan = Integer.parseInt(items[4]);
-        rowWeight = Integer.parseInt(items[5]);
-        columnWeight = Integer.parseInt(items[6]);
+        rowWeight = Float.parseFloat(items[5]);
+        columnWeight = Float.parseFloat(items[6]);
     }
 }
